requirements.txt
boto==2.20.1
celery==3.1.7
docker-py==0.3.0
Django==1.5.1
django-bootstrap3-datetimepicker==2.2.3
django-celery==3.1.1
django-celery-email==1.0.4
django-compressor==1.3
django-debug-toolbar==1.0.1
django-extensions==1.2.5
django-model-utils==2.0.3
django-storages==1.1.8
django-widget-tweaks==1.3
djangorestframework==2.3.10
djrill==0.7.0
fabric==1.8.3
numpy==1.8.0
Pillow==2.3.1
psycopg2==2.5.1
pycrypto==2.6.1
PyPDF2==1.19
redis==2.9.0
requests==2.2.1
South==0.8.4
wsgiref==0.1.2
